[+] Refactored the new signal feature
