+++ includes/sighandler.h
@@ -1,0 +1,7 @@
+#ifndef SIGHANDLER_H
+# define SIGHANDLER_H
+
+void handle_sigint(int signal);
+void handle_sigalarm(int signal);
+
+#endif
